move drawer and update tests
