+++ packages/website/src/routes/Drones/index.test.tsx
@@ -16,6 +16,9 @@ describe("Drones", () => {
         loading: false,
         error: null,
       },
+      reefsList: {
+        list: [],
+      },
     });
 
     store.dispatch = jest.fn();
